Revert "Linear ticker snapshot and delta and merged stream"
