+++ Cargo.toml
@@ -34,7 +34,6 @@ rand = "0.9.1"
 futures = "0.3.25"
 thiserror = "2.0.12"
 log = "0.4.27"
-async-stream = "0.3"
 
 [dev-dependencies]
 tokio = { version = "1", features = ["full", "test-util"] }
