--- src/models/linear_ticker/mod.rs
@@ -1,8 +1,0 @@
-#[allow(clippy::module_inception)]
-mod linear_ticker_data;
-mod linear_ticker_delta;
-mod linear_ticker_snapshot;
-
-pub use linear_ticker_data::*;
-pub use linear_ticker_delta::*;
-pub use linear_ticker_snapshot::*;
