+++ src/models/linear_ticker_data.rs
@@ -1,9 +1,13 @@
 use crate::prelude::*;
 
-#[derive(Clone, PartialEq, Deserialize, Serialize)]
+/// Structure for linear perpetual futures ticker data.
+///
+/// Contains ticker metrics specific to linear perpetual futures, such as funding rates and open interest. Bots use this for real-time market analysis and risk management in USDT-margined contracts.
+
+#[derive(Deserialize, Serialize, Clone)]
 #[serde(rename_all = "camelCase")]
 #[serde(deny_unknown_fields)]
-pub struct LinearTickerDataDelta {
+pub struct LinearTickerData {
     /// The trading pair symbol (e.g., "BTCUSDT").
     ///
     /// The ONLY required field for the ticker data. Bots use this to identify the market they are trading in.
@@ -14,7 +18,7 @@ pub struct LinearTickerDataDelta {
     /// The tick direction of the last price change.
     ///
     /// Indicates whether the last price change was an uptick or downtick (e.g., "PlusTick"). Bots use this to analyze short-term price momentum.
-    pub tick_direction: Option<TickDirection>,
+    pub tick_direction: Option<String>,
 
     /// The 24-hour price change percentage.
     ///
@@ -146,9 +150,9 @@ pub struct LinearTickerDataDelta {
     pub open_interest: Option<f64>,
 }
 
-impl std::fmt::Debug for LinearTickerDataDelta {
+impl std::fmt::Debug for LinearTickerData {
     fn fmt(&self, fmt: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let mut debug_struct = fmt.debug_struct("LinearTickerDataDelta");
+        let mut debug_struct = fmt.debug_struct("LinearTickerData");
 
         debug_struct.field("symbol", &self.symbol);
 
@@ -220,5 +224,25 @@ impl std::fmt::Debug for LinearTickerDataDelta {
         }
 
         debug_struct.finish()
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use itertools::Itertools;
+
+    use crate::fixture;
+
+    use super::*;
+
+    #[test]
+    fn deserialize() {
+        let json = fixture!("ws_linear_ticker");
+        let values = serde_json::from_str::<Vec<WsTicker>>(json)
+            .unwrap()
+            .into_iter()
+            .map(|t| t.data.try_unwrap_linear().unwrap())
+            .collect_vec();
+        assert_eq!(values.len(), 102);
     }
 }
