+++ src/models/mod.rs
@@ -64,7 +64,7 @@ mod leverage_filter;
 mod leverage_request;
 mod liability_qty;
 mod liability_qty_data;
-mod linear_ticker;
+mod linear_ticker_data;
 mod liquidation;
 mod liquidation_data;
 mod long_short_ratio;
@@ -133,7 +133,6 @@ mod spot_ticker_data;
 mod subscription;
 mod switch_list;
 mod switch_list_data;
-mod tick_direction;
 mod ticker;
 mod ticker_data;
 mod tickers_info;
@@ -224,7 +223,7 @@ pub use leverage_filter::*;
 pub use leverage_request::*;
 pub use liability_qty::*;
 pub use liability_qty_data::*;
-pub use linear_ticker::*;
+pub use linear_ticker_data::*;
 pub use liquidation::*;
 pub use liquidation_data::*;
 pub use long_short_ratio::*;
@@ -293,7 +292,6 @@ pub use spot_ticker_data::*;
 pub use subscription::*;
 pub use switch_list::*;
 pub use switch_list_data::*;
-pub use tick_direction::*;
 pub use ticker::*;
 pub use ticker_data::*;
 pub use tickers_info::*;
